@@ -1,9 +1,15 @@
 <template>
 	<div class="mb-4 row">
-		<label :for="id" class="col-sm-4 col-form-label pt-0 pt-sm-1">
-			{{ label }}
+		<label
+			:for="id"
+			:class="`col-sm-${leftWidth} col-form-label d-sm-flex align-items-sm-baseline`"
+		>
+			<div class="w-100">
+				<span class="label">{{ label }}</span>
+				<small v-if="optional" class="d-sm-block ms-2 ms-sm-0">optional</small>
+			</div>
 		</label>
-		<div class="col-sm-8">
+		<div :class="`col-sm-${rightWidth}`">
 			<slot />
 		</div>
 	</div>
@@ -15,6 +21,23 @@ export default {
 	props: {
 		id: String,
 		label: String,
+		optional: Boolean,
+		smallValue: Boolean,
+	},
+	computed: {
+		leftWidth() {
+			return this.smallValue ? "6" : "4";
+		},
+		rightWidth() {
+			return this.smallValue ? "6" : "8";
+		},
 	},
 };
 </script>
+<style scoped>
+.label {
+	-webkit-hyphens: auto;
+	hyphens: auto;
+	max-width: 100%;
+}
+</style>
