@@ -7,15 +7,15 @@ import (
 	"time"
 
 	"github.com/evcc-io/evcc/api"
+	"github.com/evcc-io/evcc/util/config"
 	"github.com/spf13/cobra"
 )
-
-// cmdEcho.AddCommand
 
 // chargerRampCmd represents the charger command
 var chargerRampCmd = &cobra.Command{
 	Use:   "ramp [name]",
 	Short: "Ramp current from 6..16A in configurable steps",
+	Args:  cobra.MaximumNArgs(1),
 	Run:   runChargerRamp,
 }
 
@@ -76,22 +76,12 @@ func runChargerRamp(cmd *cobra.Command, args []string) {
 	}
 
 	// select single charger
-	if err := selectByName(cmd, &conf.Chargers); err != nil {
+	if err := selectByName(args, &conf.Chargers); err != nil {
 		log.FATAL.Fatal(err)
 	}
 
-	if err := cp.configureChargers(conf); err != nil {
+	if err := configureChargers(conf.Chargers); err != nil {
 		log.FATAL.Fatal(err)
-	}
-
-	chargers := cp.chargers
-	if len(args) == 1 {
-		name := args[0]
-		charger, err := cp.Charger(name)
-		if err != nil {
-			log.FATAL.Fatal(err)
-		}
-		chargers = map[string]api.Charger{name: charger}
 	}
 
 	digits, err := strconv.Atoi(cmd.Flags().Lookup(flagDigits).Value.String())
@@ -104,11 +94,13 @@ func runChargerRamp(cmd *cobra.Command, args []string) {
 		log.ERROR.Fatalln(err)
 	}
 
-	for _, c := range chargers {
-		if _, ok := c.(api.ChargerEx); digits > 0 && !ok {
+	chargers := config.Chargers().Devices()
+
+	for _, v := range config.Instances(chargers) {
+		if _, ok := v.(api.ChargerEx); digits > 0 && !ok {
 			log.ERROR.Fatalln("charger does not support mA control")
 		}
-		ramp(c, digits, delay)
+		ramp(v, digits, delay)
 	}
 
 	// wait for shutdown
