@@ -3,9 +3,9 @@ package configure
 import (
 	"errors"
 	"fmt"
+	"slices"
 
 	"github.com/evcc-io/evcc/util/templates"
-	"golang.org/x/exp/slices"
 )
 
 // configureDeviceGuidedSetup lets the user choose a device that is set to support guided setup
