@@ -1,6 +1,8 @@
 package configure
 
-import "github.com/evcc-io/evcc/util/templates"
+import (
+	"github.com/evcc-io/evcc/util/templates"
+)
 
 const (
 	DefaultConfigFilename string = "evcc.yaml"
