@@ -11,7 +11,7 @@ import (
 //go:embed demo.yaml
 var demoYaml string
 
-func demoConfig(conf *config) error {
+func demoConfig(conf *globalConfig) error {
 	viper.SetConfigType("yaml")
 	if err := viper.ReadConfig(strings.NewReader(demoYaml)); err != nil {
 		return fmt.Errorf("failed decoding demo config: %w", err)
