@@ -11,14 +11,15 @@ import (
 
 	"github.com/evcc-io/evcc/api"
 	"github.com/evcc-io/evcc/core/coordinator"
-	"github.com/evcc-io/evcc/core/db"
 	"github.com/evcc-io/evcc/core/loadpoint"
 	"github.com/evcc-io/evcc/core/planner"
+	"github.com/evcc-io/evcc/core/session"
 	"github.com/evcc-io/evcc/core/soc"
 	"github.com/evcc-io/evcc/core/wrapper"
 	"github.com/evcc-io/evcc/provider"
 	"github.com/evcc-io/evcc/push"
 	"github.com/evcc-io/evcc/util"
+	"github.com/evcc-io/evcc/util/config"
 	"github.com/keep94/sunrise"
 
 	evbus "github.com/asaskevich/EventBus"
@@ -176,14 +177,14 @@ type Loadpoint struct {
 	progress                *Progress      // Step-wise progress indicator
 
 	// session log
-	db      db.Database
-	session *db.Session
+	db      session.Database
+	session *session.Session
 
 	tasks *util.Queue[Task] // tasks to be executed
 }
 
 // NewLoadpointFromConfig creates a new loadpoint
-func NewLoadpointFromConfig(log *util.Logger, cp configProvider, other map[string]interface{}) (*Loadpoint, error) {
+func NewLoadpointFromConfig(log *util.Logger, other map[string]interface{}) (*Loadpoint, error) {
 	lp := NewLoadpoint(log)
 	if err := util.DecodeOther(other, lp); err != nil {
 		return nil, err
@@ -221,18 +222,20 @@ func NewLoadpointFromConfig(log *util.Logger, cp configProvider, other map[strin
 	lp.collectDefaults()
 
 	if lp.MeterRef != "" {
-		var err error
-		if lp.chargeMeter, err = cp.Meter(lp.MeterRef); err != nil {
+		dev, err := config.Meters().ByName(lp.MeterRef)
+		if err != nil {
 			return nil, err
 		}
+		lp.chargeMeter = dev.Instance()
 	}
 
 	// default vehicle
 	if lp.VehicleRef != "" {
-		var err error
-		if lp.defaultVehicle, err = cp.Vehicle(lp.VehicleRef); err != nil {
+		dev, err := config.Vehicles().ByName(lp.VehicleRef)
+		if err != nil {
 			return nil, err
 		}
+		lp.defaultVehicle = dev.Instance()
 	}
 
 	// TODO deprecated
@@ -243,10 +246,11 @@ func NewLoadpointFromConfig(log *util.Logger, cp configProvider, other map[strin
 	if lp.ChargerRef == "" {
 		return nil, errors.New("missing charger")
 	}
-	var err error
-	if lp.charger, err = cp.Charger(lp.ChargerRef); err != nil {
+	dev, err := config.Chargers().ByName(lp.ChargerRef)
+	if err != nil {
 		return nil, err
 	}
+	lp.charger = dev.Instance()
 	lp.configureChargerType(lp.charger)
 
 	// setup fixed phases:
@@ -407,7 +411,7 @@ func (lp *Loadpoint) evChargeStartHandler() {
 	lp.socUpdated = time.Time{}
 
 	// set created when first charging session segment starts
-	lp.updateSession(func(session *db.Session) {
+	lp.updateSession(func(session *session.Session) {
 		if session.Created.IsZero() {
 			session.Created = lp.clock.Now()
 		}
@@ -553,12 +557,6 @@ func (lp *Loadpoint) applyAction(actionCfg api.ActionConfig) {
 	}
 	if max := actionCfg.MaxCurrent; max != nil && *max <= *lp.onDisconnect.MaxCurrent {
 		lp.SetMaxCurrent(*max)
-	}
-	if actionCfg.MinSoc != nil {
-		lp.SetMinSoc(*actionCfg.MinSoc)
-	}
-	if actionCfg.TargetSoc != nil {
-		lp.SetTargetSoc(*actionCfg.TargetSoc)
 	}
 	if actionCfg.Priority != nil {
 		lp.SetPriority(*actionCfg.Priority)
@@ -783,7 +781,7 @@ func (lp *Loadpoint) setStatus(status api.ChargeStatus) {
 
 // remainingChargeEnergy returns missing energy amount in kWh if vehicle has a valid energy target
 func (lp *Loadpoint) remainingChargeEnergy() (float64, bool) {
-	return math.Max(0, lp.targetEnergy-lp.getChargedEnergy()/1e3),
+	return max(0, lp.targetEnergy-lp.getChargedEnergy()/1e3),
 		!lp.vehicleHasSoc() && lp.targetEnergy > 0
 }
 
@@ -924,8 +922,8 @@ func (lp *Loadpoint) effectiveCurrent() float64 {
 
 	// adjust actual current for vehicles like Zoe where it remains below target
 	if lp.chargeCurrents != nil {
-		cur := max(lp.chargeCurrents)
-		return math.Min(cur+2.0, lp.chargeCurrent)
+		cur := max(lp.chargeCurrents[0], lp.chargeCurrents[1], lp.chargeCurrents[2])
+		return min(cur+2.0, lp.chargeCurrent)
 	}
 
 	return lp.chargeCurrent
@@ -1143,7 +1141,7 @@ func (lp *Loadpoint) pvMaxCurrent(mode api.ChargeMode, sitePower float64, batter
 	effectiveCurrent := lp.effectiveCurrent()
 	activePhases := lp.activePhases()
 	deltaCurrent := powerToCurrent(-sitePower, activePhases)
-	targetCurrent := math.Max(effectiveCurrent+deltaCurrent, 0)
+	targetCurrent := max(effectiveCurrent+deltaCurrent, 0)
 
 	lp.log.DEBUG.Printf("pv charge current: %.3gA = %.3gA + %.3gA (%.0fW @ %dp)", targetCurrent, effectiveCurrent, deltaCurrent, sitePower, activePhases)
 
@@ -1219,7 +1217,7 @@ func (lp *Loadpoint) pvMaxCurrent(mode api.ChargeMode, sitePower float64, batter
 	lp.resetPVTimer()
 
 	// cap at maximum current
-	targetCurrent = math.Min(targetCurrent, maxCurrent)
+	targetCurrent = min(targetCurrent, maxCurrent)
 
 	return targetCurrent
 }
