@@ -1,11 +1,11 @@
 package core
 
 import (
+	"slices"
 	"time"
 
 	"github.com/evcc-io/evcc/api"
 	"github.com/evcc-io/evcc/core/planner"
-	"golang.org/x/exp/slices"
 )
 
 const (
