@@ -1,7 +1,6 @@
-package db
+package session
 
 import (
-	serverdb "github.com/evcc-io/evcc/server/db"
 	"github.com/evcc-io/evcc/util"
 	"gorm.io/gorm"
 )
@@ -14,22 +13,13 @@ type DB struct {
 }
 
 type Database interface {
-	Session(startEnergy float64) *Session
+	New(startEnergy float64) *Session
 	Persist(session interface{})
 }
 
-// New creates a database storage driver
-func New(name string) (*DB, error) {
-	db := serverdb.Instance
-
-	// TODO deprecate
-	var err error
-	if table := "transactions"; db.Migrator().HasTable(table) {
-		err = db.Migrator().RenameTable(table, new(Session))
-	}
-	if err == nil {
-		err = db.AutoMigrate(new(Session))
-	}
+// NewStore creates a session store
+func NewStore(name string, db *gorm.DB) (*DB, error) {
+	err := db.AutoMigrate(new(Session))
 
 	sessiondb := &DB{
 		log:  util.NewLogger("db"),
@@ -40,8 +30,8 @@ func New(name string) (*DB, error) {
 	return sessiondb, err
 }
 
-// Session creates a charging session
-func (s *DB) Session(meter float64) *Session {
+// New creates a charging session
+func (s *DB) New(meter float64) *Session {
 	t := Session{
 		Loadpoint: s.name,
 	}
