@@ -99,7 +99,13 @@ func (s *HTTPd) RegisterSiteHandlers(site site.API, cache *util.Cache) {
 		"state":          {[]string{"GET"}, "/state", stateHandler(cache)},
 		"config":         {[]string{"GET"}, "/config/templates/{class:[a-z]+}", templatesHandler},
 		"products":       {[]string{"GET"}, "/config/products/{class:[a-z]+}", productsHandler},
-		"test":           {[]string{"POST", "OPTIONS"}, "/config/test/{class:[a-z]+}", testHandler},
+		"device":         {[]string{"GET"}, "/config/devices/{class:[a-z]+}/{id:[0-9.]+}", deviceHandler},
+		"devices":        {[]string{"GET"}, "/config/devices/{class:[a-z]+}", devicesHandler},
+		"newdevice":      {[]string{"POST", "OPTIONS"}, "/config/devices/{class:[a-z]+}", newDeviceHandler},
+		"updatedevice":   {[]string{"PUT", "OPTIONS"}, "/config/devices/{class:[a-z]+}/{id:[0-9.]+}", updateDeviceHandler},
+		"deletedevice":   {[]string{"DELETE", "OPTIONS"}, "/config/devices/{class:[a-z]+}/{id:[0-9.]+}", deleteDeviceHandler},
+		"testconfig":     {[]string{"POST", "OPTIONS"}, "/config/test/{class:[a-z]+}", testHandler},
+		"testdevice":     {[]string{"POST", "OPTIONS"}, "/config/test/{class:[a-z]+}/{id:[0-9.]+}", testHandler},
 		"buffersoc":      {[]string{"POST", "OPTIONS"}, "/buffersoc/{value:[0-9.]+}", floatHandler(site.SetBufferSoc, site.GetBufferSoc)},
 		"bufferstartsoc": {[]string{"POST", "OPTIONS"}, "/bufferstartsoc/{value:[0-9.]+}", floatHandler(site.SetBufferStartSoc, site.GetBufferStartSoc)},
 		"prioritysoc":    {[]string{"POST", "OPTIONS"}, "/prioritysoc/{value:[0-9.]+}", floatHandler(site.SetPrioritySoc, site.GetPrioritySoc)},
